+++ src/components/patterns/megalinks/components/megalink-link-list/__tests__/MegalinkLinkList.jest.spec.js
@@ -1,12 +1,13 @@
 import { config, mount } from '@vue/test-utils';
+import { setActivePinia, createPinia } from 'pinia';
 import VsMegalinkLinkList from '../MegalinkLinkList.vue';
-
-config.global.renderStubDefaultSlot = true;
 
 const videoId = '12345';
 const videoBtnText = 'Play Video';
 
-const mountConfig = {
+config.global.renderStubDefaultSlot = true;
+
+const factoryMount = (slotData) => mount(VsMegalinkLinkList, {
     propsData: {
         featured: true,
         imgSrc: 'test',
@@ -20,27 +21,23 @@ const mountConfig = {
         videoId,
         videoBtnText,
     },
-    slots: {
-        'vs-link-list-heading': 'Multi-image heading',
-        'vs-link-list-content': '<p>Multi-image content</p>',
-    },
-};
-
-const factoryMount = () => mount(VsMegalinkLinkList, mountConfig);
+    ...slotData,
+});
 
 describe('VsMegalinkLinkList', () => {
-    let wrapper;
-
     beforeEach(() => {
-        wrapper = factoryMount();
+        setActivePinia(createPinia());
     });
 
     describe(':props', () => {
         it('renders the correct theme class', () => {
+            const wrapper = factoryMount();
             expect(wrapper.find('.vs-megalink-link-list--dark').exists()).toBe(true);
         });
 
         it('should accept a `videoId` property and pass it to the stretched-link-card', () => {
+            const wrapper = factoryMount();
+
             const card = wrapper.findComponent({
                 name: 'VsStretchedLinkCard',
             });
@@ -49,6 +46,8 @@ describe('VsMegalinkLinkList', () => {
         });
 
         it('should accept a `videoBtnText` property and pass it to the stretched-link-card', () => {
+            const wrapper = factoryMount();
+
             const card = wrapper.findComponent({
                 name: 'VsStretchedLinkCard',
             });
@@ -58,13 +57,27 @@ describe('VsMegalinkLinkList', () => {
     });
     describe(':slots', () => {
         it('renders content inserted in a vs-link-list-heading slot', () => {
+            const wrapper = factoryMount({
+                slots: {
+                    'vs-link-list-heading': 'Multi-image heading',
+                },
+            });
+
             expect(wrapper.find('[data-test="megalink-link-list__title"]').text()).toBe('Multi-image heading');
         });
 
         it('renders content inserted in a vs-link-list-content slot', () => {
+            const wrapper = factoryMount({
+                slots: {
+                    'vs-link-list-content': '<p>Multi-image content</p>',
+                },
+            });
+
             expect(wrapper.find('[data-test="megalink-link-list__content"]').html()).toContain('<p>Multi-image content</p>');
         });
         it('renders card panels if days and transport are provided', () => {
+            const wrapper = factoryMount();
+
             expect(wrapper.find('[data-test="vs-itinerary-panels"]').exists()).toBe(true);
         });
     });